ml_service/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# ML and NLP libraries
torch==2.1.0
transformers==4.35.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
tokenizers==0.15.0

# Retrieval and search
rank-bm25==0.2.2
faiss-cpu==1.7.4
numpy==1.24.3
scikit-learn==1.3.2

# Summarization
sumy==0.11.0
nltk==3.8.1

# NLP utilities
spacy==3.7.2
nltk==3.8.1

# Data processing
pandas==2.1.3
pyyaml==6.0.1

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
httpx==0.25.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Optional GPU support (uncomment if needed)
# torch==2.1.0+cu118
# faiss-gpu==1.7.4

